Add unit test to components
In this commit I have added test to App and AddStoreroom components. Also, I have reorganized the components locations. We can divide components in container components and component components (when only render UI interface). In this case, App is a container component and AddStoreroom is just a component.

Furthermore, by Jest convention for testing, the tests classes go in __tests__ directory.